implementado validacoes nas rotas de categorias e desativacao de usuario

+++ src/api/controllers/categorias.js
@@ -3,6 +3,7 @@ const ErrorResponse = require('../../utils/helpers/errorResponse');
 const JsonResponse = require('../../utils/helpers/jsonResponse');
 const asyncHandler = require('../middleware/asyncHandler');
 const StatusCodes = require('http-status-codes');
+const { getCategoriaDto } = require('../../utils/helpers/categoriaHelper');
 
 class CategoriaController {
     list = asyncHandler(async (request, response, next) => {
@@ -18,20 +19,41 @@ class CategoriaController {
 
     get = asyncHandler(async (request, response, next) => {
         const categoria = await Categoria.findById(request.params.id);
+        if (!categoria) {
+            return next(
+                new ErrorResponse(
+                    'Categoria não encontrada',
+                    StatusCodes.NOT_FOUND
+                )
+            );
+        }
 
         response.status(StatusCodes.OK).json(new JsonResponse(categoria));
     });
 
     insert = asyncHandler(async (request, response, next) => {
-        const categoria = await Categoria.create(request.body);
+        const categoriaDto = getCategoriaDto(request);
+        const categoria = await Categoria.create(categoriaDto);
 
         response.status(StatusCodes.CREATED).json(new JsonResponse(categoria));
     });
 
     update = asyncHandler(async (request, response, next) => {
-        const categoria = await Categoria.findByIdAndUpdate(
+        let categoria = await Categoria.findById(request.params.id);
+        if (!categoria) {
+            return next(
+                new ErrorResponse(
+                    'Categoria não encontrada',
+                    StatusCodes.NOT_FOUND
+                )
+            );
+        }
+
+        const categoriaDto = getCategoriaDto(request);
+
+        categoria = await Categoria.findByIdAndUpdate(
             request.params.id,
-            request.body,
+            categoriaDto,
             {
                 new: true,
                 runValidators: true,
@@ -42,6 +64,16 @@ class CategoriaController {
     });
 
     delete = asyncHandler(async (request, response, next) => {
+        let categoria = await Categoria.findById(request.params.id);
+        if (!categoria) {
+            return next(
+                new ErrorResponse(
+                    'Categoria não encontrada',
+                    StatusCodes.NOT_FOUND
+                )
+            );
+        }
+
         await Categoria.deleteById(request.params.id);
 
         response.status(StatusCodes.OK).json(new JsonResponse({}));
